convert autocomplete to functional component insted of class
