+++ client/src/components/Form/Reservation/AutoComplete.js
@@ -1,47 +1,31 @@
 /* eslint-disable react/no-unused-state */
 import { AutoComplete } from 'antd';
-import React, { Component } from 'react';
+import React from 'react';
+import PropTypes from 'prop-types';
 
 const { Option } = AutoComplete;
 
-const roomsName = ['Tokyo', 'Berlin', 'Roma', 'NewYork', 'Cairo', 'Jerusalim'];
+const Autocomplete = ({ rooms, setRoom, handleSearch }) => (
+  <AutoComplete
+    style={{
+      width: 200,
+    }}
+    onSelect={setRoom}
+    onSearch={handleSearch}
+    placeholder="Room Name"
+  >
+    {rooms.map((room) => (
+      <Option key={room} value={room}>
+        {room}
+      </Option>
+    ))}
+  </AutoComplete>
+);
 
-class Autocomplete extends Component {
-  state = {
-    rooms: [],
-    selectedRoom: null,
-  };
+Autocomplete.propTypes = {
+  rooms: PropTypes.string.isRequired,
+  setRoom: PropTypes.func.isRequired,
+  handleSearch: PropTypes.func.isRequired,
+};
 
-  handleSearch = (value) => {
-    this.setState({
-      rooms: roomsName.filter((e) =>
-        e.toUpperCase().includes(value.toUpperCase())
-      ),
-    });
-  };
-
-  setRoom = (value) => {
-    this.setState({ selectedRoom: value });
-  };
-
-  render() {
-    const { rooms } = this.state;
-    return (
-      <AutoComplete
-        style={{
-          width: 200,
-        }}
-        onSelect={this.setRoom}
-        onSearch={this.handleSearch}
-        placeholder="Room Name"
-      >
-        {rooms.map((room) => (
-          <Option key={room} value={room}>
-            {room}
-          </Option>
-        ))}
-      </AutoComplete>
-    );
-  }
-}
 export default Autocomplete;
